+++ packages/synchro-charts/src/components/sc-table/sc-table-base/sc-table-base.spec.tsx
@@ -56,7 +56,7 @@ const tableBaseSpecPage = async (propOverrides: Partial<Components.ScTableBase> 
     columns: [],
     isEnabled: true,
     messageOverrides: {
-      liveModeOnlyMessage: 'liveModeOnlyMessage',
+      liveModeOnly: 'liveModeOnly',
     },
     ...propOverrides,
   };